Extracted paginator to reusable hook
