+++ frontend-react/src/pages/submissions/SubmissionTable.tsx
@@ -1,4 +1,4 @@
-import { useContext, useEffect, useState } from "react";
+import { useContext } from "react";
 import {
     Button,
     ButtonGroup,
@@ -10,8 +10,9 @@ import { NavLink } from "react-router-dom";
 
 import SubmissionsResource from "../../resources/SubmissionsResource";
 import { GlobalContext } from "../../components/GlobalContextProvider";
+import usePaginator from "../../utils/UsePaginator";
 
-import { FilterState, SubmissionFilterContext } from "./FilterContext";
+import { SubmissionFilterContext } from "./FilterContext";
 
 function SubmissionTable() {
     // this component will refresh when global context changes (e.g. organization changes)
@@ -19,54 +20,6 @@ function SubmissionTable() {
     const { filters, updateStartRange, updateEndRange, updateSortOrder } =
         useContext(SubmissionFilterContext);
 
-    /*    Pagination behavior
-     * 1. Load page. Start range filter is used to dictate first cursor
-     * 2. On response, store cursor + nextCursor in Map<number, string> with 1-indexing to mimic page numbers
-     * 3. When updateStartRange is called, conditionally update Map with new cursor at new index
-     * 4. When paging back, use the Map to retrieve the previous cursor and set it
-     * 5. When paging forward, use the Map to retrieve the next cursor and set it
-     * 6. When calling a page, the cursor should be set to currentIndex cursor and endCursor should be set
-     *    to currentIndex + 1 cursor
-     */
-
-    // Initialized cursors with first cursor as blank so page 1 always has the newest items
-    // displayed.
-    const [cursors, updateCursors] = useState<Map<number, string>>(
-        new Map([[1, ""]])
-    );
-    const [currentIndex, updateCurrentIndex] = useState<number>(1);
-    const [hasNext, setHasNext] = useState(false);
-    const [hasPrev, setHasPrev] = useState(false);
-
-    /* NOT meant to show total page count, only page count "so far" */
-    const pageCount = () => {
-        return cursors.size;
-    };
-    const changeCursor = (cursorIndex: number) => {
-        let cursor =
-            cursors.get(cursorIndex) || cursors.get(currentIndex) || null;
-        if (cursorIndex === 1 || !cursor) {
-            /* Cursor 1 will always have the value of "" so that page 1 always has
-             * the absolute latest results pulled. However, this will return null
-             * if called cursors.get(1), so this check manually sets the cursor to
-             * an empty string.
-             *
-             * For this, we also must reset the Map so previously tracked
-             * cursors are lost and new cursors can be stored.
-             */
-            cursor = "";
-            updateCursors(new Map([[1, cursor]]));
-            updateCurrentIndex(cursorIndex);
-        } else {
-            if (cursor === cursors.get(cursorIndex))
-                updateCurrentIndex(cursorIndex);
-        }
-
-        updateStartRange!!(cursor);
-        // updateEndRange!!(endCursor)
-    };
-
-    // we can tell if we're on the first page by saving the first result and then checking against it later
     const submissions: SubmissionsResource[] = useResource(
         SubmissionsResource.list(),
         {
@@ -79,33 +32,15 @@ function SubmissionTable() {
         }
     );
 
-    // after the first page loads, set up some state info about the data.
-    useEffect(() => {
-        const cursorExists = (c: string) => {
-            return Array.from(cursors.values()).includes(c);
-        };
-        const addCursors = (nextCursor: string) => {
-            updateCursors(cursors.set(currentIndex + 1, nextCursor));
-        };
-        const updatePrevNextBools = () => {
-            setHasNext(currentIndex < cursors.size);
-            setHasPrev(currentIndex > 1);
-        };
-
-        const lastTimestamp =
-            submissions[filters.pageSize - 1]?.createdAt || null;
-        if (lastTimestamp && !cursorExists(lastTimestamp))
-            addCursors(lastTimestamp);
-        updatePrevNextBools();
-
-        debugger;
-    }, [
-        currentIndex,
-        updateCurrentIndex,
-        cursors,
-        submissions,
-        filters.pageSize,
-    ]);
+    /* Gives handlers for all pagination needs!
+     *
+     * - hasPrev/hasNext: boolean, indicating previous and next pages exist.
+     * - currentIndex: number, the page number you're currently on.
+     * - changeCursor: function(desiredCursorIndex), handles cursor navigation and updating.
+     * - pageCount: function(), returns the current size of your cursor Map
+     */
+    const { hasPrev, hasNext, currentIndex, changeCursor, pageCount } =
+        usePaginator(submissions, filters, updateStartRange!!);
 
     const getSortedSubmissions = (): SubmissionsResource[] => {
         submissions?.sort((a, b) => SubmissionsResource.sortByCreatedAt(a, b));
